adding validations to erease posts and styling

+++ src/components/ReplyCard/ReplyCard.jsx
@@ -1,20 +1,29 @@
-import React, { useState } from "react";
+import React, { useContext } from "react";
 import { Image } from "react-bootstrap";
 import CommunityService from './../../services/community.services';
 
 import './ReplyCard.css'
+import { AuthContext } from "../../context/auth.context";
 
-import * as Icon from 'react-bootstrap-icons'
 
 const ReplyCard = ({ reply, onDeleteReply }) => {
 
+    const { user } = useContext(AuthContext)
+
     const deleteReply = () => {
-        CommunityService.deleteReply(reply._id)
-            .then(() => {
-                onDeleteReply(reply._id);
-            })
-            .catch((err) => console.log(err));
-    };
+        if (user && reply.owner._id === user._id) {
+            CommunityService.deleteReply(reply._id)
+                .then(() => {
+                    onDeleteReply(reply._id);
+                })
+                .catch((err) => console.log(err));
+        }
+    }
+
+    const renderDeleteButton = user && reply.owner._id === user._id && (
+        <p onClick={deleteReply} className="btn-delete-icon ms-auto">Eliminar</p>
+    );
+
 
     return (
         <li className='d-flex post-replies-list'>
@@ -22,9 +31,7 @@ const ReplyCard = ({ reply, onDeleteReply }) => {
             <div className='d-flex flex-column post-reply-section'>
                 <strong className='post-reply-username mb-1'>{reply.owner.username}</strong>
                 <p className="mb-0">{reply.reply}</p>
-                <p
-                    onClick={deleteReply}
-                    className="btn-delete-icon ms-auto">Eliminar</p>
+                {renderDeleteButton}
             </div>
         </li>
     );
